feat: 봇별 랭킹 조회 (STEP 10, 읽기 전용)

+++ src/main/kotlin/com/example/baseball/domain/user/UserRepository.kt
@@ -2,5 +2,6 @@ package com.example.baseball.domain.user
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserRepository : JpaRepository<User, Int> {
+// User.id 는 Long PK 이므로 JpaRepository 의 ID 타입도 Long 이어야 한다.
+interface UserRepository : JpaRepository<User, Long> {
 }
